+++ Obsolete/ChocolateyProvider.sln
@@ -1,0 +1,29 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2013
+VisualStudioVersion = 12.0.31101.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ChocolateyProvider", "ChocolateyProvider.csproj", "{12BC926F-9911-47DB-A490-21E8686514F6}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{D4933EC0-5563-4BE5-88D3-FA1ADC0BE218}"
+	ProjectSection(SolutionItems) = preProject
+		.nuget\NuGet.Config = .nuget\NuGet.Config
+		.nuget\NuGet.exe = .nuget\NuGet.exe
+		.nuget\NuGet.targets = .nuget\NuGet.targets
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{12BC926F-9911-47DB-A490-21E8686514F6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{12BC926F-9911-47DB-A490-21E8686514F6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{12BC926F-9911-47DB-A490-21E8686514F6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{12BC926F-9911-47DB-A490-21E8686514F6}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
