+++ Obsolete/packages.config
@@ -1,0 +1,7 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="chocolatey.lib" version="0.9.10.3" targetFramework="net40" />
+  <package id="log4net" version="2.0.3" targetFramework="net40" />
+  <package id="Microsoft.Web.Xdt" version="2.1.1" targetFramework="net40" />
+  <package id="NuGet.Core" version="2.8.2" targetFramework="net40" />
+</packages>
