+++ Obsolete/provider.manifest
